BL0 is updated using SMC353's ECC wrapper
SMC353's ECC wrapper is used to decode the content from NVRAM.
However, the details of the SMC353's ECC wrapper is not known, yet.
We made the placeholder of the SMC353's ECC wrapper in Qemu.
The addresses used in the code must be updated once the hardware
details are known.