trch,rtps: tests: mmu: define the address macros
So that tests build. Not sure if allocation is safe or not.
Tests disabled by default because conflict with RTPS's TEST_RT_MMU (RTPS's half
of the TRCH's TEST_RT_MMU_BASIC test). The swap test does pass (orthogonal
to the conflict). The 32-bit access test does not not pass, not sure what
is going wrong.